+++ crates/checker/src/expressions/path/variable.rs
@@ -41,6 +41,10 @@ fn visit_chain_segment(
     let variable_prop = find_variable_prop(&mut visitor.tc, host_ty, &generic_args, member_name);
     if let Some((sym, ty)) = variable_prop {
         visitor.read(sym, segment.loc);
+        if !visitor.tc.is_visible(sym.into()) {
+            let error = DiagnosticKind::FieldIsPrivate(member_name.to_string());
+            visitor.error(error, segment.ident.loc);
+        }
         return Some(ir::Expression::Member(ir::MemberExpression {
             loc: Location::merge(host.loc(), segment.loc),
             object: Some(Box::new(host)),
@@ -51,6 +55,10 @@ fn visit_chain_segment(
 
     if let Some((sym, ty)) = visitor.tc.find_method(host.ty(), member_name, |_| true) {
         visitor.read(sym, segment.loc);
+        if !visitor.tc.is_visible(sym.into()) {
+            let error = DiagnosticKind::MethodIsPrivate(member_name.to_string());
+            visitor.error(error, segment.ident.loc);
+        }
         return Some(ir::Expression::Method(ir::MethodExpression {
             loc: Location::merge(host.loc(), segment.loc),
             host: Some(Box::new(host)),