+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Fixed
 
+- Fixed field visibility on path expressions
 - Fixed public items import from other modules
 - Fixed dom element codegen
 - Fixed markup coloring